+++ app/gui/src/project-view/components/ComponentMenu.vue
